fixed/files imports, error input state
