+++ src/assets/images/mobile/bg-register-mob.png
