+++ src/components/RestrictedRoute/RestrictedRoute.jsx
@@ -7,7 +7,6 @@ export default function RestrictedRoute({
   redirectTo = '/',
 }) {
   const isLoggedIn = useSelector(state => state.auth.isLoggedIn);
-  console.log(isLoggedIn);
 
   return isLoggedIn ? <Navigate to={redirectTo} /> : <Component />;
